+++ tests/unit/customizations/test_cloudsearchdomain.py
@@ -12,6 +12,7 @@
 # language governing permissions and limitations under the License.
 from awscli.testutils import unittest
 from awscli.testutils import BaseAWSCommandParamsTest
+from awscli.help import HelpRenderer
 from awscli.customizations.cloudsearchdomain import validate_endpoint_url
 
 import mock
@@ -48,14 +49,24 @@ class TestSearchCommand(BaseAWSCommandParamsTest):
         stderr = self.run_cmd(cmd, expected_rc=255)[1]
         self.assertIn('--endpoint-url is required', stderr)
 
+    def test_endpoint_not_required_for_help(self):
+        cmd = self.prefix + 'help'
+        with mock.patch('awscli.help.get_renderer') as get_renderer:
+            mock_render = mock.Mock(spec=HelpRenderer)
+            get_renderer.return_value = mock_render
+            stdout, stderr, rc = self.run_cmd(cmd, expected_rc=None)
+            # If we get this far we've succeeded, but we can do
+            # a quick sanity check and make sure the service name is
+            # in the stdout help text.
+            self.assertIn(stdout, 'cloudsearchdomain')
+
 
 class TestCloudsearchDomainHandler(unittest.TestCase):
     def test_validate_endpoint_url_is_none(self):
-        parsed_args = mock.Mock()
-        parsed_args.endpoint_url = None
-        parsed_args.command = 'cloudsearchdomain'
+        parsed_globals = mock.Mock()
+        parsed_globals.endpoint_url = None
         with self.assertRaises(ValueError):
-            validate_endpoint_url(parsed_args)
+            validate_endpoint_url(parsed_globals)
 
 
 if __name__ == "__main__":